+++ Applications/KNN/jni/meson.build
@@ -1,13 +1,13 @@
 res_path = join_paths(meson.current_source_dir(), '..', 'res')
 
 knn_sources = [
-  'main.cpp',
+  'main_sample.cpp',
   'bitmap_helpers.cpp'
 ]
 
 knn_inc = include_directories('.')
 
-e = executable('nntrainer_knn',
+e = executable('knn_sample',
   knn_sources,
   dependencies: [iniparser_dep, nntrainer_dep, tflite_dep],
   install: get_option('install-app'),
